Med Card section implementation

+++ data/src/main/java/kg/iaau/diploma/data/AccessToken.kt
@@ -1,8 +1,8 @@
 package kg.iaau.diploma.data
 
 data class AccessToken(
-    var id: Int? = null,
-    var userId: Int? = null,
+    var id: Long? = null,
+    var userId: Long? = null,
     var accessToken: String? = null,
     var tokenExpirationTime: String? = null,
     var refreshToken: String? = null,
